Rename and move to utils package
